refactor: style frame instead & set body to full height
