+++ .storybook/config.js
@@ -1,13 +1,15 @@
 import { configure, addDecorator } from '@storybook/react'
 import { jsxDecorator } from 'storybook-addon-jsx'
 import { withPropsTable } from 'storybook-addon-react-docgen'
-import { CssResetWrapper } from './reset-decorator'
+import { CssResetWrapper } from './css-reset-decorator'
+import { FrameStylesWrapper } from './frame-styles-decorator'
 
 import 'typeface-roboto'
 
 addDecorator(jsxDecorator)
 addDecorator(withPropsTable)
 addDecorator(CssResetWrapper)
+addDecorator(FrameStylesWrapper)
 
 function loadStories() {
     const req = require.context('../stories', true, /\.stories\.js$/)
