+++ .storybook/css-reset-decorator.js
@@ -1,12 +1,12 @@
-import React from 'react'
+import React, { Fragment } from 'react'
 
 import { CssReset } from '../src'
 
 const CssResetWrapper = fn => (
-    <div style={{ margin: 16 }}>
+    <Fragment>
         <CssReset />
         {fn()}
-    </div>
+    </Fragment>
 )
 
 export { CssResetWrapper }
